// src/types/founder.ts
export interface FounderProps {
  name: string;
  role: string;
  description: string;
  longDescription: string;
  expertise: string[];
  profileImage: string;
  yearsExperience: number;
}

export interface FounderProfile {
  name: string;
  role: string;
  longDescription: string;
  expertise: string[];
  profileImage: string;
  yearsExperience: number;
}